requirements.txt
selenium==4.5.0
python-docx==0.8.11
docx2pdf==0.1.8
tk